add uidl and top support, clear c99 warnings

+++ pop3filter/src/parse_helpers.c
@@ -15,7 +15,7 @@
  * If args = 0. Accept all commands.
  */
 char ** sctp_parse_cmd(buffer *b, struct management *data, int *args, int *st_err){
-    int argc = POPG_ARGC_BLOCK;
+    size_t argc = POPG_ARGC_BLOCK;
     char ** cmd = malloc(argc * sizeof(char *));
 
     if (cmd == NULL){
